+++ Metropolis/src/utopia/metropolis/model/partial/user/UserSettingsData.scala
@@ -9,10 +9,10 @@ import java.time.Instant
 
 object UserSettingsData extends FromModelFactoryWithSchema[UserSettingsData]
 {
-	override def schema = ModelDeclaration("name" -> StringType, "email" -> StringType)
+	override def schema = ModelDeclaration("name" -> StringType)
 	
-	override protected def fromValidatedModel(model: Model[Constant]) = UserSettingsData(model("name").getString,
-		model("email").getString)
+	override protected def fromValidatedModel(model: Model[Constant]) =
+		UserSettingsData(model("name").getString, model("email").string)
 }
 
 /**
@@ -20,10 +20,20 @@ object UserSettingsData extends FromModelFactoryWithSchema[UserSettingsData]
   * @author Mikko Hilpinen
   * @since 2.5.2020, v1
   * @param name user-name
-  * @param email User's email address
+  * @param email User's email address (if specified)
  *  @param created Creation time of this settings version
   */
-case class UserSettingsData(name: String, email: String, created: Instant = Now) extends ModelConvertible
+case class UserSettingsData(name: String, email: Option[String] = None, created: Instant = Now) extends ModelConvertible
 {
+	// COMPUTED -------------------------------------
+	
+	/**
+	  * @return Whether this data contains an email address
+	  */
+	def specifiesEmail = email.nonEmpty
+	
+	
+	// IMPLEMENTED  ---------------------------------
+	
 	override def toModel = Model(Vector("name" -> name, "email" -> email, "created" -> created))
 }