+++ Metropolis/src/utopia/metropolis/model/stored/user/UserSettings.scala
@@ -9,11 +9,10 @@ import utopia.metropolis.model.stored.Stored
 
 object UserSettings extends FromModelFactoryWithSchema[UserSettings]
 {
-	override val schema = ModelDeclaration("id" -> IntType, "user_id" -> IntType, "name" -> StringType,
-		"email" -> StringType)
+	override val schema = ModelDeclaration("id" -> IntType, "user_id" -> IntType, "name" -> StringType)
 	
 	override protected def fromValidatedModel(model: Model[Constant]) = UserSettings(model("id").getInt,
-		model("user_id").getInt, UserSettingsData(model("name").getString, model("email").getString,
+		model("user_id").getInt, UserSettingsData(model("name").getString, model("email").string,
 			model("created").getInstant))
 }